% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_color}
\alias{average_color}
\title{Get the average hex color from a list of colors}
\usage{
average_color(color_names)
}
\arguments{
\item{color_names}{Vector of color names (from [grDevices::colors()]).}
}
\value{
Average color in hex form.
}
\description{
First the average hue is calculated by averaging r, g and b components of the
colors and converted to HSL. Then, this hue is combined with average s and l
in order to avoid muddy colors. Finally, the combined HSV is converted to
hex.
}
