% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question.R
\name{create_question}
\alias{create_question}
\title{Create an HTML file with the quiz question}
\usage{
create_question(n, quiz = getOption("peRson.quiz"))
}
\arguments{
\item{n}{Sequence number of the quiz question.}

\item{quiz}{Quiz environment with quiz variables (uses questions,
question_colors, named_colors, css_file).}
}
\value{

}
\description{
Creates an HTML file with the question, answers, answer images, image holders
for results and leaderboard and navigation to the previous and next question.
}
