% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_setup.R
\name{create_quiz_pages}
\alias{create_quiz_pages}
\title{Create all the pages of the quiz}
\usage{
create_quiz_pages(quiz = getOption("peRson.quiz"))
}
\arguments{
\item{quiz}{Quiz environment with quiz variables.}
}
\value{
Writes quiz pages quiz/Q{{n}}.html from 0 to number of questions plus
  1.
}
\description{
Create the intro, questions and favourite question pages in one step.
}
