% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_sheets.R
\name{create_summary_sheet}
\alias{create_summary_sheet}
\title{Create a summary answer sheet}
\usage{
create_summary_sheet(participants, n_questions)
}
\arguments{
\item{participants}{Character vector with the names of the participants.}

\item{n_questions}{Number of questions in the quiz.}
}
\value{
The id of the sheet as a character string.
}
\description{
The function creates a Google Sheet on Google Drive containing links to the
participants' answer sheets (created with [create_answer_sheet()]). The sheet
contains one column per participant and n_questions + 1 rows (additional row
is for the favourite question). Import range functions are inserted into the
first row, however, you need to go to the file and manually link the sheets
by clicking on cells with REF!.
}
