% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{demo_answers}
\alias{demo_answers}
\title{Demo quiz answers}
\format{
A data frame with 22 rows and 7 variables:
\describe{
  \item{Beverly}{Beverly's answers}
  \item{Caty}{Caty's answers}
  \item{Doug}{Doug's answers}
  \item{Drake}{Drake's answers}
  \item{Fawn}{Fawn's answers}
  \item{Spike}{Spike's answers}
  \item{Teddy}{Teddy's answers}
}
}
\usage{
demo_answers
}
\description{
A dataset containing answers provided by fixional quiz participants.
}
\keyword{datasets}
