% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{demo_favourite}
\alias{demo_favourite}
\title{Demo quiz favourite question}
\format{
A data frame with 1 row and 7 variables:
\describe{
  \item{Beverly}{Beverly's favourite question}
  \item{Caty}{Caty's favourite question}
  \item{Doug}{Doug's favourite question}
  \item{Drake}{Drake's favourite question}
  \item{Fawn}{Fawn's favourite question}
  \item{Spike}{Spike's favourite question}
  \item{Teddy}{Teddy's favourite question}
}
}
\usage{
demo_favourite
}
\description{
A dataset containing the favourite question provided by fixional quiz participants.
}
\keyword{datasets}
