% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{demo_participants}
\alias{demo_participants}
\title{Demo quiz participants}
\format{
A data frame with 7 rows and 4 variables:
\describe{
  \item{name}{name of the participant}
  \item{color}{chosen color of the participant}
  \item{image}{path to the participant's image}
  \item{email}{participant's email}
}
}
\usage{
demo_participants
}
\description{
A dataset containing data for fixional quiz participants.
}
\keyword{datasets}
