% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{demo_questions}
\alias{demo_questions}
\title{Demo quiz questions}
\format{
A data frame with 21 rows and 10 variables:
\describe{
  \item{person}{name of the participant who provided the question}
  \item{text}{question}
  \item{answer_A}{first answer choice}
  \item{answer_B}{second answer choice}
  \item{answer_C}{third answer choice}
  \item{answer_D}{fourth answer choice}
  \item{image_A}{path to the image for the first answer choice}
  \item{image_B}{path to the image for the first answer choice}
  \item{image_C}{path to the image for the first answer choice}
  \item{image_D}{path to the image for the first answer choice}
}
}
\usage{
demo_questions
}
\description{
A dataset containing questions provided by fixional quiz participants.
}
\keyword{datasets}
