% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_answers.R
\name{evaluate_answers}
\alias{evaluate_answers}
\title{Evaluate the correctness of question answers}
\usage{
evaluate_answers(
  answers = NULL,
  n = NULL,
  correct_answer = NULL,
  quiz = quiz.env
)
}
\arguments{
\item{answers}{A data frame with the answers (see [demo_answers] for format).
If NULL, answers are read from the quiz summary sheet (summary_sheet_id)
created by [quiz_setup()].}

\item{n}{Sequence number of the question to evaluate. If NULL, the next
question based on the length of saved answers in [quiz.env] is evaluated.}

\item{correct_answer}{Correct answer to the question. If NULL, the answer of
the person who asked the question is taken as the correct answer.}

\item{quiz}{Quiz environment with quiz variables (uses summary_sheet_id,
answers, questions).}
}
\value{
Evaluated answers are saved to quiz.env answers list. A plot with
  results is printed and saved to "quiz/A{n}.png".
}
\description{
Evaluate the correctness of question answers
}
