% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourite.R
\name{favourite_question}
\alias{favourite_question}
\title{Create an HTML file with the favourite question question}
\usage{
favourite_question(n = NULL, quiz = getOption("peRson.quiz"))
}
\arguments{
\item{n}{Number of questions to include in the table.}

\item{quiz}{Quiz environment with quiz variables (uses answers (length),
questions (n, person, text)).}
}
\value{
Writes quiz/Q{{n+1}}.html file.
}
\description{
Creates an HTML file with the table of asked questions and their sequence
numbers, so participants can vote about their favourite question. Also
contains navigation to the previous and next question.
}
