% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourite.R
\name{favourite_result}
\alias{favourite_result}
\title{Create an HTML file with the favourite question answer}
\usage{
favourite_result(answer, n = NULL, quiz = quiz.env)
}
\arguments{
\item{answer}{Data frame with the answer. Must contain one column for each
participant named with their name and one row with their answers (see
[demo_favourite]).}

\item{n}{Number of questions in the quiz (excluding the favourite questions
question).}

\item{quiz}{Quiz environment with quiz variables (uses answers (length),
questions (n, person, text)).}
}
\value{

}
\description{
Creates an HTML file with the table of questions that were chosen as
favourites, ordered by number of votes. Also contains navigation to the
previous and next question.
}
\details{
As the results of the favourite question are processed in this function, it
also saves favourite_result table to the quiz environment.
}
