% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourite.R
\name{favourite_table}
\alias{favourite_table}
\title{Create an HTML favourite question table}
\usage{
favourite_table(questions, n, quiz = quiz.env)
}
\arguments{
\item{questions}{Data frame with questions to include in the table (uses n,
person, text)}

\item{n}{Page/question number}

\item{quiz}{Quiz environment with quiz variables (uses css_file,
question_colors, named_colors).}
}
\value{

}
\description{
Create an HTML favourite question table
}
