% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_sheets.R
\name{fill_demo_answer_sheets}
\alias{fill_demo_answer_sheets}
\title{Fill demo answer sheets with answers for testing purposes}
\usage{
fill_demo_answer_sheets(quiz = quiz.env, answers = demo_answers)
}
\arguments{
\item{quiz}{Quiz environment with quiz variables (uses participants).}

\item{answers}{Answers to write into the sheets (see [demo_answers] for
format).}
}
\value{
Fills the answer sheets created by [quiz_setup()] with answers, so
  you can test [evaluate_answers()].
}
\description{
Fill demo answer sheets with answers for testing purposes
}
