% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_results.R
\name{final_results}
\alias{final_results}
\title{Create an HTML file with final quiz results}
\usage{
final_results(quiz = getOption("peRson.quiz"))
}
\arguments{
\item{quiz}{Quiz environment with quiz variables (uses questions, answers,
presence, all_questions, favourite_results, participants, named_colors,
css_file)}
}
\value{
Writes quiz/Q{{n_questions+3}}.html file.
}
\description{
Computes the final results and creates an HTML file with participants'
images, their colors and badges they have earned.
}
