% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color_dist}
\alias{get_color_dist}
\title{Calculate distances between colors in RGB space}
\usage{
get_color_dist(r, g, b)
}
\arguments{
\item{r}{Vector of values for red.}

\item{g}{Vector of values for green.}

\item{b}{Vector of values for blue.}
}
\value{
Vector containing the average distances of colors from other colors
in the vector.
}
\description{
Calculate distances between colors in RGB space
}
