% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo_quiz.R
\name{get_demo_participants}
\alias{get_demo_participants}
\title{Get demo participants data frame}
\usage{
get_demo_participants()
}
\value{
A data frame with demo participant and correct paths to their images.
}
\description{
Get demo participants data frame
}
