% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_bing_image}
\alias{get_first_bing_image}
\title{Get the URL of the first image search result on Bing}
\usage{
get_first_bing_image(search_term)
}
\arguments{
\item{search_term}{The search term to use in image search. Only
alphanumerical characters and spaces will be included.}
}
\value{
Character string with the URL of the first image result or path to
  transparent image if there are no search results.
}
\description{
Google makes it hard to scrape images, so Bing it is. :D
}
