% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_question_color}
\alias{get_question_color}
\title{Get the color for the question based on its sequence number}
\usage{
get_question_color(n, question_colors, named_colors)
}
\arguments{
\item{n}{The sequence number of the question.}

\item{question_colors}{List of question colors as created by [quiz_setup()].}

\item{named_colors}{Named list of colors as created by [quiz_setup()].}
}
\value{
The color of the question for sequence numbers within questions and
  average color otherwise.
}
\description{
Helper function for assembling the navigation bars.
}
