% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transparent_pic}
\alias{get_transparent_pic}
\title{Get the correct system path to transparent picture}
\usage{
get_transparent_pic()
}
\value{
System path to the transparent picture.
}
\description{
Transparent picture is used as a placeholder for quiz answers and
leaderboard. It is supplied with the package, but needs to be accessed with
[base::system.file()].
}
