% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header_div}
\alias{header_div}
\title{HTML code for header div}
\usage{
header_div(text, bgcolor)
}
\arguments{
\item{text}{Text to include in the header.}

\item{bgcolor}{Background color of the header (CSS compatible string).}
}
\value{
HTML div element
}
\description{
HTML code for header div
}
