% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question.R
\name{preview_images}
\alias{preview_images}
\title{Create an HTML file with all quiz questions for image preview}
\usage{
preview_images(questions)
}
\arguments{
\item{questions}{A data frame with the questions (see [demo_questions] for format)}
}
\value{
Writes quiz/preview_images.html file.
}
\description{
Creates an HTML file with the questions, answers and answer images. This way
you can preview all images, replace some if necessary and have them loaded in
the browser cache, so they load faster during the quiz.
}
