% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_setup.R
\name{process_colors}
\alias{process_colors}
\title{Process colors chosen by participants}
\usage{
process_colors(participants)
}
\arguments{
\item{participants}{A data frame with participants (see [demo_participants]
for format).}
}
\value{
A data frame with random colors chosen for participants who didn't
  choose colors, r, g, b columns with RGB values, hex with hex color value
  and dist with the average distance from other participants' colors.
}
\description{
For participants who didn't choose colors a random color is picked. Colors
are converted to RGB and hex to use in HTML files. Distance from other
participants' colors is calculated for final results.
}
