% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{quiz.env}
\alias{quiz.env}
\title{Quiz environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
quiz.env
}
\description{
An empty environment that will contain the variables created by quiz generating functions [quiz_setup()], [evaluate_answers()], [favourite_answer()].
}
\keyword{datasets}
