% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_setup.R
\name{quiz_setup}
\alias{quiz_setup}
\title{Set up the quiz environment}
\usage{
quiz_setup(
  questions,
  participants,
  presence = NULL,
  shuffle = TRUE,
  create_sheets = TRUE
)
}
\arguments{
\item{questions}{A data frame with questions (see [demo_questions] for
format).}

\item{participants}{A data frame with participants (see [demo_participants]
for format).}

\item{presence}{Vector of present participants' names. If NULL, all
participants from participants data frame (participants$name) will be
deemed present.}

\item{shuffle}{If TRUE, the order of answers and questions is randomly
shuffled by [shuffle_answers()] and [shuffle_questions()] respectively.}

\item{create_sheets}{If TRUE, a Google Sheet for answers is created for and
shared with each participant, plus one summary sheet containing imported
answer sheets (though the sheets need to be linked manually in the sheet
before answers can be read from it).}
}
\value{
A quiz environment with the following variables. \describe{
  \item{presence}{vector or present participants' names} \item{css_file}{path
  to css file used to style the HTML quiz files} \item{all_questions}{data
  frame with all the provided questions from the questions data frame
  (shuffled if shuffle is TRUE)} \item{questions}{data frame with questions
  that will be used in the quiz (only those whose askers are present)}
  \item{participants}{data frame with participants from the participants data
  frame (with added variables chose_color, dist, present and answer_sheet if
  create_sheets is TRUE)} \item{summary_sheet_id}{if create_sheets is TRUE
  character id of the created summary sheet} \item{named_colors}{vector of
  hex colors chosen by participants, named with their names, plus average
  color of chosen colors with "avg" name} \item{question_colors}{vector of
  hex colors for each question in questions based on the person who asked it}
  \item{answers}{empty list that will be filled with answers by
  [evaluate_answers()]} }
}
\description{
Set up the quiz environment
}
