% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_color}
\alias{random_color}
\title{Select a random color}
\usage{
random_color(exclude_colors = NA)
}
\arguments{
\item{exclude_colors}{Vector of color names (from [base::colors()]) to
exclude from random selection.}
}
\value{
A random color name
}
\description{
Excludes gray and white colors and colors from exclude_colors list and
returns one random color.
}
