% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo_quiz.R
\name{run_demo_quiz}
\alias{run_demo_quiz}
\title{Run demo quiz}
\usage{
run_demo_quiz(
  questions = demo_questions,
  participants = get_demo_participants(),
  answers = demo_answers,
  favourite = demo_favourite,
  create_sheets = FALSE
)
}
\arguments{
\item{questions}{A data frame with questions (see [demo_questions] for
format).}

\item{participants}{A data frame with participants (see [demo_participants]
for format).}

\item{answers}{A data frame with answers (see [demo_answers] for format).}

\item{favourite}{A data frame with favourite question (see [demo_favourite]
for format).}

\item{create_sheets}{If TRUE, sheets for answers will be created on Google
Drive as they would be in a real quiz and trashed at the end.}
}
\value{
Creates quiz HTML files and results images.
}
\description{
Setup the quiz, create questions, evaluate answers, favourite question and
final results with demo data.
}
