% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_contestants.R
\name{show_contestants}
\alias{show_contestants}
\title{Create the quiz intro HTML page with participants' images and colors}
\usage{
show_contestants(presence = NULL, n_per_row = 5, quiz = quiz.env)
}
\arguments{
\item{presence}{Vector of present participants. If NULL, taken from quiz
environment (presence).}

\item{n_per_row}{Number of pictures per row.}

\item{quiz}{Quiz environment with quiz variables (uses presence,
named_colors, question_colors, css_file, participants)}
}
\value{
Saves Q0.html file.
}
\description{
Create the quiz intro HTML page with participants' images and colors
}
