% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_answers.R
\name{show_leaderboard}
\alias{show_leaderboard}
\title{Create the current leaderboard}
\usage{
show_leaderboard(answers = NULL, n = NULL, quiz = quiz.env)
}
\arguments{
\item{answers}{Current list of evaluated answers. If NULL, taken from quiz
environment.}

\item{n}{Current question. If NULL, estimated based on the length of saved
answers in quiz environment.}

\item{quiz}{Quiz environment with quiz variables (uses answers, named_colors,
presence)}
}
\value{
A plot with leaderboard is printed and saved to "quiz/L{n}.png".
}
\description{
Create the current leaderboard
}
