% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_setup.R
\name{shuffle_answers}
\alias{shuffle_answers}
\title{Randomly assign answers to A-D}
\usage{
shuffle_answers(questions)
}
\arguments{
\item{questions}{A data frame with questions (see [demo_questions] for
format).}
}
\value{
A data frame with shuffled answers.
}
\description{
The answers to questions are randomly assigned to letters A-D.
}
