% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_questions.R
\name{shuffle_questions}
\alias{shuffle_questions}
\title{Randomly shuffle questions in groups}
\usage{
shuffle_questions(questions, shuffle_by = rn)
}
\arguments{
\item{questions}{data frame containing quiz questions and a rn variable that represents the random sequence number of the questions that were provided by the same participant}
}
\value{
data frame with shuffled questions
}
\description{
Divides the questions into groups containing one question per participant, then randomly shuffles the order of the questions within the group. This way the order of the questions will be random, but a question from the same participant won't appear again until a question from every participant is used.
}
