% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_sheets.R
\name{quiz_clean_drive}
\alias{quiz_clean_drive}
\title{Deletes the answer sheets}
\usage{
quiz_clean_drive(
  sheet_ids = c(quiz.env$participants$answer_sheet, quiz.env$summary_sheet_id)
)
}
\arguments{
\item{sheet_ids}{Vector of string sheet ids.}
}
\value{
No result, just a drive cleaned of quiz traces.
}
\description{
Puts the participants' answer sheets and the summary answer sheet into drive trash.
}
