% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_sheets.R
\name{quiz_clean_drive}
\alias{quiz_clean_drive}
\title{Deletes the answer sheets}
\usage{
quiz_clean_drive(quiz = getOption("peRson.quiz"))
}
\arguments{
\item{quiz}{Quiz environment with quiz variables.}
}
\value{
No result, just trashes the quiz sheets.
}
\description{
Puts the participants' answer sheets and the summary answer sheet into drive trash.
}
