% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_setup.R
\name{shuffle_questions}
\alias{shuffle_questions}
\title{Randomly shuffle questions in rounds}
\usage{
shuffle_questions(questions, shuffle_by = rn)
}
\arguments{
\item{questions}{A data frame with quiz questions (see [demo_questions] for
format) and a rn variable that represents the random sequence number of the
questions that were provided by the same participant.}

\item{shuffle_by}{Unquoted name of the variable used to create shuffling
groups.}
}
\value{
A data frame with shuffled questions.
}
\description{
Divides the questions into rounds containing one question per participant,
then randomly shuffles the order of the questions within the round. This way
the order of the questions will be random, but a question from the same
participant won't appear again until a question from every participant is
used.
}
